use `ddi_prop_get_int` not `ddi_getprop`
The latter is obsolete, and predates the typed DDI property interfaces.

This unfortunately means on devicetree systems such as arm, where devicetree
is defined to store big-endian data but the CPU is little-endian, only the
typed interfaces have enough information to byteswap, and uses of the old
interfaces will potentially (if a request reaches the "PROM"), return data in
network order.