use typed `ddi_prop_lookup...` in cases that might affect aarch64
As with `ddi_getprop` `ddi_getlongprop` is obsolete, and will return
big-endian data for integer properties sourced from the PROM, stop using it as
much as possible, mark the functions as obselete on aarch64, and gag the warning
from that precisely.

This should mean that as well as noticing the cases here, we catch any future
uses that might give surprising results on FDT-based aarch64.