Simplify boot UART and reset interaction
The boot UART is presently a hardcoded, platform specific chunk of code
that is needed very early in boot. This all prevents us from introducing
platmods for platform-specifics.

Refactor bootsvcs UART and reset interactions into a separate file and
drive the selection of the UART implementation from xboot_info. This
means that inetboot (already deeply platform specific) is now
responsible for setting up the correct parameters for the UART and for
performing any initialisation required, as the bsvc code in the kernel
only knows how to send and receive characters.