Use loaders to abstract the arch_timer frequency
We have a very early FDT dependency in unix for the arch_timer frequency. This
dependency exists to deal with the case where the firmware has not programmed
the CNTFRQ_EL0 register, requiring a fallback to firmware tables.

The loaders can't simply ensure that this value is properly programmed, as it
can only be written from the highest implemented EL, which is going to be EL3
on most proper hardware (i.e. where ATF runs).

Since we already have this information in the loaders, pass the corrected value
through to unix using xboot_info. To ensure that we're setting this as early as
reasonable (without repurposing kobj_start any more than we already have), pass
xboot_info as a second argument to mlsetup to allow setting the frequency after
the initial module linkage and immediately before the timer functions are
selected.