Make PROM usable on both FDT and ACPI systems
Most of the PROM, other than node and prop functions, is the same as what is
found on i86pc, and therefore the same between FDT and ACPI systems. Make the
PROM itself usable on both, supporting runtime selection of the firmware
interface and making it possible to move the bulk of the PROM back to genunix
at a later stage.

Note that our FDT PROM includes a lot of helper functions. These gave been
renamed to have a prom_fdt_ prefix, making it clear that they are not generic
PROM functions. When built in DEBUG the FDT implementation asserts that the
FDT pointer is set when any user-callable PROM functions are called. This is
intended to help non-FDT porting efforts and turn hard-to-debug crashes into
obvious errors.

Since we've cleaned up prom_init.c and now have the same compilation guards as
intel/i86pc, delete the custom prom_init.c from kmdb.

Note that I do little more than what's absolutely necessary to clean up the
PSM PROM implementation. I'd like to see this disappear or be radically
minimised in the future.