Make PROM usable on both FDT and ACPI systems
Most of the PROM, other than node and prop functions, is the same as what is
found on i86pc, and therefore the same between FDT and AC{I systems. Make the
PROM itself usable on both, supporting runtime selection of the firmware
interface and making it possible to move the bulk of the PROM back to genunix
at some stage.

Note that our FDT PROM includes a lot of helper functions. These gave been
renamed to have a prom_fdt_ prefix, making it clear that they are not generic
PROM functions.

Const-correctness has also been improved throughout the PROM interfaces, making
it easier to use the functions when (for example) the name one os searching for
is const.