Hacky PCIe prototype
This began life with two goals, which less overlap than it seems (which is
part of the problem)

1. The DDI layers need a lot of work to support real hardware properly, of
   which PCIe is probably the most complicated we care about right now, so
   seemed the obvious target to shake out such issues.

2. PCIe support is just, well, necessary.  And even on hardware where it's
   optional (Raspberry Pi 4x) it would allow for much better quality of
   life (fast storage!)

So this is a PCIe prototype intended to try to do both parts (which, again,
wasn't a good idea).

The prototype is structured like this, for ease of prototyping:

- We support one, ECAM, root complex, which we expect at /pcie in the device
  tree (this is what qemu gives us).

- We don't support the full PCI Firmware spec booting/initialization, we
  expect firmware to have done initialization (this requires
  richlowe/arm64-gate's qemu-uboot branch).

- We handle interrupts in a manner more like x86, even though the structure of
  x86, with the PSM etc, isn't a great way to go about this.

- We only support FIXED interrupts (because it would make more sense to
  implement MSI only after working out the interrupt architecture, above)

The prototype is full of comments, either XXXARM, XXXROOTNEX, or XXXPCI (not
necessarily organized as sensibly as this implies).  Many of them are super
important things I punted on completely).

The prototype is full of a rather unfortunate idiom of
`ASSERT(0 && "some text")`
as a sort of "not implemented"/"unreachable" macro.

This is especially relevant regarding interrupts, where many of the interrupt
operations are not necessary in the tight constraints of the prototype.

The implementation of interrupt mapping is distasteful, but sort of arranged
to make it easier (for me) to debug as I was working on making interrupt stuff
generally better.

I haven't made anything `pbchk` clean yet, overlong lines, cstyle complaining,
lions and tigers!

I'm running and testing basic prototype-y things by copying
`qemu-setup/run.sh` and adding

```
     -bios u-boot.bin \
     -device pcie-root-port,id=rp20,bus=pcie.0,chassis=1,addr=2.0 \
     -netdev user,id=user0 -device igb,mac=52:54:98:76:54:32,bus=rp20,netdev=user0,id=igb0 \
     "$@"
```

This will give you an igb0 on the bus, which you can configure with `ipadm
create-addr -Tdhcp igb0/dhcp`, and then use as per usual qemu user networking.

virtio pci doesn't yet work (or attach).  I didn't debug that yet.

Things to do:

- Every XXX* comment (some of these are questions, some are bugs, some are
  questions that imply bugs).

- My hope is that interrupts get structured such that the normal DDI flow
  applies to them _except_ that we hand `i_ddi_intr_ops` each dominating
  interrupt domain all the back up to the interurpt controller (rather than
  the devicetree parent).  This implies that gicv2 and gicv3 have drivers
  which bind to the nodes in the tree, and who present interrupt operations.

  I have a 2nd prototype of this, but there are a further chunk of issues that
  come to the surface when we do this.

- My hope is that `pci_autoconfig` as a module goes away, and we can have
  `npe` attach, and `pcicfg_configure` its children as if they were
  hotplugged.  This would mean we only had one path to configure PCI busses,
  that it was the better, more complete path, and that we no longer relied on
  every bus the system sees having been correctly configured by firmware.

- Multiple root complexes, and non-ECAM config space need us to hang the
  config space address and method off the root complex dip in some manner and
  look it up later (we may also need to be able to do this based solely on
  bdf, without a dip?)