Resurrect parent private data to simplify resource configuration
Resource configuration, i.e. registers, ranges and interrupts, is currently
overly complicated in the aarch64 port, and open-coded throughout the
codebase. For some reason, parent private data has been completely ignored
in the base port work.

To simplify interacting with resources, move all reasoning about parsing
resources to the `make_ddi_ppd` path, refactoring the logic to split out
reasoning about register, ranges and interrupts. Extend the base parent
private data object to include information about the interrupt configuration
(currently only sense information) and use DDI constants
(`DDI_INTR_FLAG_EDGE` and `DDI_INTR_FLAG_LEVEL`) to express this information
in a firmware-independent manner.

This, of course, means that all busses should use the
`impl_ddi_sunbus_initchild` and `impl_ddi_sunbus_removechild` helpers or,
at a minimum, the `make_ddi_ppd` and `impl_free_ddi_ppd` helpers. Busses
with wildly differing resource formats (ACPI, for example), will need to
convert resources or remap them as needed.

An interesting divergence from dtspec is that in dtspec an empty ranges
property indicates an identity mapping and absence of a range property
indicates that no translation is possible across this boundary. This is
not somehting that can be expressed in parent private data, where we
assume that no translation entries means identity mapping. This is not
as much of a problem as it seems, as the affected nexus will include
code that understands that no mapping is available and understands how
to address and interact with child devices.

Another interesting item that came up is that the root node was missing
the hardware properties expressed in the FDT. Specifically these are
`interrupt-parent`, `#interrupt-cells`, `#address-cells`, `#size-cells`
and `dma-coherent`. We now copy these properties into the root node at
initialisation time to ensure that all of our resource-handling logic
hangs together (this was not a problem before, as raw PROM access was
used to access the FDT, but we now use DDI interfaces).

The `ranges` is not copied over from the FDT, even through we've kept
the ranges-handling code around in the rootnex. Ranges don't really
make sense on rootnex, as there's no parent address information (since
there's no parent).

Resource handling on the root nexus has been reworked to interact with
the parent private data only. In particular this affects the `bus_map`
and `intr_ops` functionlity.

For `bus_map` the changes include an update from i86pc to perform
regspec64 conversion on any regspec request that gets to the root nexus.
The mapping code itself uses the device arena now, rather that simply
figuring out where in `kpm` the `inetboot.bin` loader may (or may not)
have set up the device mapping. This part sounds surprising, but this is
how it was working - the loader set up the direct-map mappings for
known I/O devices only and the kernel simply returned those for mapping
requests - bus faults would occur when attaching a device that the loader
had not set up (such as the reworked GIC, further up this stack).

For `intr_ops` the changes significantly simplify the handler by using
parent private data rather than open-coded FDT queries. Interrupts that
do not have configuration expressed (none, in FDT) are simply left as-is.

All regspec and rangespec interaction mentioned above is performed via a
set of macros that understand how to pack a 64-bit address into a 32-bit
regspec. There are verify statements dotted around the macros to ensure
that we know about data that cannot be handled, and the macros support
a physical address space of up to 56 bits, which is the same as the
architecture supports. Sizes are limited to 32-bit quantities, so 4G. If
this becomes of a problem we can tackle a full regspec64 migration
alongside i86pc.

With the reworked rootnex in place most of what's in `simple-bus` becomes
unnecessary, since we are now aware of parent private data from the root
nexus, and what was in simple-bus was just more of the same open-coded
FDT query logic. `simple-bus` has been significantly simplified, calling
`i_ddi_bus_map` and `i_ddi_intr_ops` for resource handling, making it
simply in nature as well as in name.

With all of the above in place it became necessary to slightly rework
the `viommionex` to copy it's resource-affecting properties to its
child node, leaving enough information in place for the standard resource
parsing code to do the right thing.

Handling of the interrupt tree is still nowhere close to perfect, and needs
the work Richard has been chipping away at to make it more robust, but at
least it's all in one place now.