remove `i_ddi_get_inum` which isn't useful and needs work to implement
This returned the vector number given an index into its "interrupts", but this
requires the knowledge of the bus binding, or an operation to ask the bus
nexus for the information.

Since we don't use this call/information anymore, stop using it.  A reason to
bring this back would be that `ih_vector` needs to be kept up-to-date for
compatibility.