gicthree: attach as a driver and implemented interrupt ops
This is, because of how things work, exactly the gictwo code right now, which
is a shame but also makes sense.

While here somewhat clarify the (shared) problem with shared interrupts, now
that some exist (via PCIe).

While here clarify that quiesce is not _unnecessary_ for GIC, it's _not
supported_.  We would need to reload it from software.