Make aarch64 cbe a proper driver
Implement an Arm Generic Timer driver to take over the duties of
the cyclic backend. Probe this driver from the existing `cbe_init`
function.

Being a proper driver means that we no longer need to know about
the firmware-specific details of IRQs and can attach the interrupt
parent on demand. Doing so allows us to make the GIC drivers be
full DDI-drivers.