rpi4: Fix reading CPU speed via the mailbox interface
There was a bit of confusion here, where the call into the clock rate
reading code expects an FDT clock identifier, but the CPU speed code was
passing a mailbox identifier. Correct this to pass the FDT clock ID and
handle the FDT clock ID, translating it into a mailbox clock ID alongside
the other supported clocks.