aarch64: Move virt modules to the armv8 directory
The only module in the virt machine is `todpl031`, which is standalone,
represents real hardware and is not named the way we'd prefer (spelling out
a trailing digit).

Move this driver to `armv8` and rename it to `pl03one`. Hook it up to the
build in the new location using a simplified makefile with no warning gags.

There's still some work that needs to go in here to load this driver from
the device tree - this is tricky because the TOD is needed very early.