virt: Move todpl031 to armv8 and rename
The `todpl031` module is not virt-specific - it represents real
hardware, described in DDI0224 Arm PrimeCell Real Time Clock (PL031).
The module is also not named the way we'd prefer (spelling out the
trailing digit).

Move this driver to `armv8` and rename it to `pl03one`. Hook it up to
the build in the new location using a simplified makefile with no
warning gags.

Ideally this driver would attach properly and use DDI for resource
management, but this is particularly hard given how early the TOD
is needed, so I've left this for another time.