ddi: rename `sparc_pd_...` -> `i_ddi_pd_...`
This is more truthful -- they're far from SPARC specific.

However, I suspect the intent of the parent private data is, unsurprisingly
that it's private to the parent, not to the DDI (the whole dev_info is private
to the DDI...).  This reinforces that not being true.

I have left them implemented as macros, though that isn't necessary in any way
I can determine.