gtmr: Use the correct interrupt when running at EL2
When `unix` runs at EL2 in unified EL12 mode we don't fire the physical
timer interrupt, but rather the Hypervisor interrupt (this despite
interacting with the physical timer registers). Recognise this setup
in the architected timer `cbe` driver and wire up the right interrupts.

There are no changes when running at EL1 (as we do everywhere at the
moment), but with the next change landed the following works:

```
QEMU_SCRIPT_CPU="neoverse-n2" \
  QEMU_SCRIPT_MACHINE="virt,gic-version=3,virtualization=on" \
  QEMU_SCRIPT_NCPU=4 ./run.sh
```

Without this change the above command hangs after hooking up the cyclic
backend (in main). The hang is in `callout_init`, and caused by `cbe`
not receiving the expected interrupts.